ML Pipeline/requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
scikit-learn
joblib
python-dotenv